Advanced merge
