+++ qrutils/mathUtils/mathUtils.pri
@@ -2,10 +2,8 @@ HEADERS += \
 	$$PWD/math.h \
 	$$PWD/gaussNoise.h \
 	$$PWD/geometry.h \
-	$$PWD/mathUtils.h \
 
 SOURCES += \
 	$$PWD/math.cpp \
 	$$PWD/gaussNoise.cpp \
 	$$PWD/geometry.cpp \
-	$$PWD/mathUtils.cpp \
